PIMCORE-1876 - Do not allow "admin", "webservice", "install" and "plugin" as name for documents in the first level (system routes)
git-svn-id: http://www.pimcore.org/svn/pimcore/private/core/trunk@2465 1f8fe7d8-47f0-464c-8d0a-336f4953ab05

+++ pimcore/models/Document.php
@@ -411,6 +411,11 @@ class Document extends Pimcore_Model_Abstract implements Document_Interface {
             throw new Exception("Document requires key, document with id " . $this->getId() . " deleted");
         }
 
+        $disallowedKeysInFirstLevel = array("install","admin","webservice","plugin");
+        if(in_array($this->getKey(), $disallowedKeysInFirstLevel)) {
+            throw new Exception("Key: " . $this->getKey() . " is not allowed in first level (root-level)");
+        }
+
         // save properties
         $this->getProperties();
         $this->getResource()->deleteAllProperties();
